refactor: Major refactor in profile section  - renamed folder  - moved all form actions into a seperate folder  - cleaned up imports  - moved validations into a seperate folder  - abstracted reusable utils and component specific utils  - docs to api callers
