+++ Web Frontend/src/components/EditUserProfile/PhotoPreview.jsx
@@ -6,13 +6,13 @@ import { convertToPixelCrop } from "react-image-crop";
 import "react-image-crop/dist/ReactCrop.css";
 import toast from "react-hot-toast";
 
-import setCanvasPreview from "./setCanvasPreview";
+import setCanvasPreview from "@/lib/setCanvasPreview";
 import { compressImage } from "@/lib/compressImage";
 import { cn } from "@/lib/utils";
 import { isValidImageSize, generateCenteredCrop } from "@/lib/imageUtils";
 import userIconWhite from "@/assets/user_icon_white.png";
-import SimpleButton from "../ui/SimpleButton";
-import OutlinedButton from "../ui/OutlinedButton";
+import SimpleButton from "@/components/ui/SimpleButton";
+import OutlinedButton from "@/components/ui/OutlinedButton";
 
 import { updateAvatarApi } from "@/services/profileService";
