+++ Web Frontend/src/components/EditUserProfile/SiblingsInfo.jsx
@@ -1,73 +1,27 @@
 import { useActionState, useState } from "react";
-import toast from "react-hot-toast";
 
 import useExpandable from "@/hooks/useExpandable";
 
 import Input from "@/components/ui/Input";
-import SimpleButton from "../ui/SimpleButton";
-import OutlinedButton from "../ui/OutlinedButton";
-import SelectOption from "../ui/SelectOption";
-import Checkbox from "../ui/CheckBox";
+import SimpleButton from "@/components/ui/SimpleButton";
+import OutlinedButton from "@/components/ui/OutlinedButton";
+import SelectOption from "@/components/ui/SelectOption";
+import Checkbox from "@/components/ui/CheckBox";
 import SiblingsTable from "./SiblingTable";
 
+import siblingInfoAction from "./form_actions/siblingInfoAction";
 import {
   CLASSES,
   DIVISIONS,
   SIBLING_RELATIONS,
-} from "./helpers/siblingsInfoMaps";
-import { validateSiblingForm } from "./helpers/siblingValidation";
-
-import { addSiblingInfoApi } from "@/services/studentInfoService";
-
-const siblingFormAction = async (prevState, formData, setInfo) => {
-  const values = {
-    relation: formData.get("relation") || "Brother",
-    name: formData.get("name") || "",
-    age: Number(formData.get("age") || 1),
-    isInSameSchool: formData.get("isInSameSchool") === "on",
-    class: formData.get("class") || "Jr. KG",
-    div: formData.get("div") || "A",
-  };
-
-  try {
-    validateSiblingForm(values);
-
-    const apiCall = () => addSiblingInfoApi(values);
-    const response = await toast.promise(apiCall, {
-      loading: "Saving...",
-      success: "Sibling information saved successfully!",
-      error: "Failed to save sibling information.",
-    });
-
-    if (response.statusCode < 200 || response.statusCode > 299) {
-      throw new Error(response.error || "Something went wrong while updating.");
-    }
-
-    setInfo(response.student.siblingInfo);
-    return {
-      relation: "Brother",
-      name: "",
-      age: 1,
-      isInSameSchool: false,
-      class: "Jr. KG",
-      div: "A",
-    };
-  } catch (err) {
-    const message =
-      err?.response?.data?.message ||
-      err?.message ||
-      "An unknown error occurred while updating your profile.";
-    toast.error(message);
-    return values;
-  }
-};
+} from "./value_maps/siblingsInfoMaps";
 
 const SiblingsInfo = ({ initialInfo }) => {
   const { height, setExpanded, containerRef, expanded } = useExpandable(false);
   const [info, setInfo] = useState(initialInfo);
 
   const [formValues, formAction, isPending] = useActionState(
-    (prevState, formData) => siblingFormAction(prevState, formData, setInfo),
+    (prevState, formData) => siblingInfoAction(prevState, formData, setInfo),
     {
       relation: "Brother",
       name: "",
