+++ Web Frontend/src/components/EditUserProfile/value_maps/parentContantMaps.js
@@ -1,0 +1,4 @@
+export const PARENT_RELATIONS = [
+  { value: "Father", text: "Father" },
+  { value: "Mother", text: "Mother" },
+];
