+++ Web Frontend/src/components/EditUserProfile/value_maps/siblingsInfoMaps.js
@@ -18,11 +18,6 @@ export const SIBLING_RELATIONS = [
   { value: "Sister", text: "Sister" },
 ];
 
-export const PARENT_RELATIONS = [
-  { value: "Father", text: "Father" },
-  { value: "Mother", text: "Mother" },
-];
-
 export const DIVISIONS = [
   { value: "A", text: "A" },
   { value: "B", text: "B" },
